bug fixes for conversion
length was being incremented for every character read from stream.
Should only restrict to reference extensions.  Counter was taken
out in lieu of using the reference string directly.  The reference
string shouldn't be kept (as with the alt string when it agrees with
ref) as it's wasteful.

Diploid streams still a wip (not working, untested).

Testing still needs to be done.